rutuja_binance_bot_project/requirements.txt
python-binance==1.0.16
reportlab
pytest
